usermanagement/requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
PyJWT==2.7.0
SQLAlchemy==2.0.18
